# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kopi Tiang Alam API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [public]
      summary: List all active products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [public]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
  /orders:
    post:
      operationId: createOrder
      tags: [public]
      summary: Create a new order (dine-in or takeaway)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{orderNumber}:
    get:
      operationId: getOrderByNumber
      tags: [public]
      summary: Get order status by order number
      parameters:
        - name: orderNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  # ─────────────────── ADMIN AUTH ───────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required: [id, name, price, status, createdAt]
      properties:
        id:
          type: integer
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        image:
          type: ["string", "null"]
        stock:
          type: integer
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
...
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, price, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImage:
          type: ["string", "null"]
        quantity:
          type: integer
        price:
          type: number
        subtotal:
          type: number
    OrderItemInput:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
    Order:
      type: object
      required: [id, orderNumber, customerName, orderType, status, totalPrice, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        customerName:
          type: string
        tableNumber:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [dine_in, takeaway]
        status:
          type: string
          enum: [pending, processing, preparing, ready, completed, cancelled]
        notes:
          type: ["string", "null"]
        totalPrice:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
    OrderInput:
      type: object
      required: [customerName, orderType, items]
      properties:
        customerName:
          type: string
          minLength: 1
        tableNumber:
          type: ["string", "null"]
        orderType:
          type: string
          enum: [dine_in, takeaway]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
...