# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kopi Tiang Alam API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputOrderType } from "./orderInputOrderType";
import type { OrderItemInput } from "./orderItemInput";

export interface OrderInput {
  /** @minLength 1 */
  customerName: string;
  /** @nullable */
  tableNumber?: string | null;
  orderType: OrderInputOrderType;
  /** @nullable */
  notes?: string | null;
  items: OrderItemInput[];
}